==> code/0-globals.R <==
# Global parameters

# Paths to data files / folders

PATH_DB_RAW <- "/Volumes/ExtremeSSD/marketcheck/db/clean/db"

PATH_DB <- here::here("data", "depreciation.parquet")
PATH_DB_FOLDER <- here::here("data", "depreciation")

# Min and max vehicle ages

AGE_YEARS_MIN <- 1
AGE_YEARS_MAX <- 8
